Ingest historical bwauth statistics data
Update the parseOldConsensuses.py script to ingest historical
bwauth data. Note that this script is generally run one time
and is not intended to be a well-maintained script for future use.
It needs care and feeding for each major run.

Additionally create a mergeDatabase script. This script...
probably does not do what we want. It's included mostly as a
placeholder for future development if we want to correct and use
it in the future.