Update the bwauth data on each new consensus
This is a squashed commit that contained the following bugfixes:

 - Do not attempt to backfill table before checking if it exists

In "Create database placeholders" we didn't check that the table existed first.

 - In python int > None, which meant that we were recording all non-measurements as 'below' values

In the increment if/else statement for counting how many of
each type (median, below, etc), a comparison of int > None
returns true, which led to the described.

We changed the line to
if r.is_unmeasured or vote.routers[r.fingerprint].measured == None:
  continue

 - This should accurately track unmeasured routers in bwauth votes. I hope.

That change did not count any routers as unmeasured. So we changed it to

if r.is_unmeasured:
  continue
elif r.fingerprint not in vote.routers or vote.routers[r.fingerprint].measured == None:
  data[dirauth_nickname]['unmeasured'] += 1