Make sure that dirauths keep a consistent color between graphs.
Closes #12

This works by assigning a named css class to a dirauth as well as an unnamed one.
unnamed classes may move around based on dirauth ordering, but named classes
will override the numbered one.

When a new dirauth is added, it is likely that they will wind up sharing a color
with a named dirauth (unless they're the last dirauth in the list), but
dirauth additions are infrequent enough that I ought to be able to add a new
CSS class for them in short order.